+++ src/geometry/vertex.cpp
@@ -1,0 +1,6 @@
+#include "coli/geometry/vertex.h"
+
+namespace Coli::Geometry {
+    template COLI_EXPORT Vertex<false>;
+    template COLI_EXPORT Vertex<true>;
+}