+++ CHANGES.md
@@ -13,6 +13,9 @@
 
 - Graphics:
   - New `Vertex Array` resource
+- Geometry:
+  - New namespace `Coli::Geometry` for geometry features
+  - New `Vertex` struct and its traits `VertexTraits`
 
 ### v0.2.3